Remove cuttlefish_kernel
Move the TEST_MAPPING file so this extra presubmit testing triggers
if the cuttlefish kernel modules are touched, just as before.

History is not preserved like it was for cuttlefish_common because
the history of the cuttlefish_kernel project is not that interesting.

Add Steve to OWNERS for GKI kernel updates.

Change-Id: Id2d3ddecd8632ca10c9d28edc1c6a2f8cebf2d4f