cuttlefish_host_only -> cuttlefish_buildhost_only
The term (1) "host" in the context of cuttlefish (and more broadly in
the context of VMs) means the OS that directly talks to the hardware and
is responsible for managing the (guest) VMs. However, in the context of
the Android build system, the same term (2) "host" means the OS that
isn't Android (which is referred to as "target"), like the Ubuntu OS
where the build system itself runs on, or the Windows or Mac OSes that
we cross-build SDKs for.

Previously when cuttlefish doesn't support nested virtualization, the
two different "host"s meant the same; host-side components (like crosvm,
run_cvd, etc.) were only built for the build-host OS (Ubuntu or gLinux).

However, that is no longer true as we aim to support running cuttlefish
(and other OS payloads as well) on Android devices as well. Then the
host-side VM managing components have to be built for the target
(Android) OS.

In order to better reflect the new situation, cuttlefish_* cc_defaults
modules are renamed as follows:

1) cuttlefish_guest_only: for the guest OS (i.e. cuttlefish)
2) cuttlefish_host: for the host OSes (either Ubuntu or Android)
This was previously "cuttlefish_host_and_guest", where "guest" there
actually meant the Android OS which doesn't necessarily need to be
cuttlefish.
3) cuttlefish_buildhost_only: for non-Android host OSes (i.e. Ubuntu)

Bug: 167675429
Test: mma under /device/google/cuttlefish
Change-Id: I2d6619bed8ad95ba88c4c07b78fe844daa8f40fd