Change fake microdroid bootloader to avbpubkey
The subsequent processing of the bootloader by microdroid requires that
the bootloader contains the avbpubkey and that the file is at least a
few byte large. In order to remove the multiple workarounds in the
microdroid build rules, use the avbpubkey as the fake bootloader as it
satisfies the requirements.

Test: x86 build on forrest
Bug: 234564414
Change-Id: I731e25207a86068ff003516c3ca8b57edd196e26