IKA: Update kernel prebuilts for x86_64
Newest x86_64 AL kernel seems to boot on Ika (yay!), so this is a good
time to get back into making pick_prebuilt work. However, it's been a
long time since we updated Ika prebuilts, so some modules have been
deleted, which causes problems with pick_prebuilt's initial commit for
boring technical reasons.

(The boring technical reasons: pick_prebuilt checks the ab/ build used
for each kernel, and for each module, before and after landing new
kernels. But the modules currently on ToT weren't built from ab/; they
were built on a workstation (mine). So after pick_prebuilt downloads new
modules, it tries to read the ab/ version of each module. For modules
that are now deleted, we obviously don't download new artifacts, and
pick_prebuilt doesn't delete old artifacts, so we still try to read the
ab/ version of the old module... which is impossible, so it raises an
error.)

So anyway, I downloaded these prebuilts from ab/12741783, and this will
make it easier for us to actually start using pick_prebuilt.

Bug: 372541004
Test: Build cf_x86_64_desktop with new kernel prebuilts. Boot and check
      kernel version.
Change-Id: I92089c6970d5011fd15019ce25411f55522d6941